+++ instrumentation/opentelemetry-instrumentation-celery/tests/test_tasks.py
@@ -20,11 +20,22 @@ from wrapt import wrap_function_wrapper
 from opentelemetry import baggage, context
 from opentelemetry.instrumentation.celery import CeleryInstrumentor, utils
 from opentelemetry.instrumentation.utils import unwrap
+from opentelemetry.semconv.attributes.exception_attributes import (
+    EXCEPTION_MESSAGE,
+    EXCEPTION_STACKTRACE,
+    EXCEPTION_TYPE,
+)
 from opentelemetry.semconv.trace import SpanAttributes
 from opentelemetry.test.test_base import TestBase
 from opentelemetry.trace import SpanKind, StatusCode
 
-from .celery_test_tasks import app, task_add, task_raises, task_returns_baggage
+from .celery_test_tasks import (
+    CustomError,
+    app,
+    task_add,
+    task_raises,
+    task_returns_baggage,
+)
 
 
 class TestCeleryInstrumentation(TestBase):
@@ -127,15 +138,15 @@ class TestCeleryInstrumentation(TestBase):
         self.assertEqual(1, len(consumer.events))
         event = consumer.events[0]
 
-        self.assertIn(SpanAttributes.EXCEPTION_STACKTRACE, event.attributes)
+        self.assertIn(EXCEPTION_STACKTRACE, event.attributes)
 
-        # TODO: use plain assertEqual after 1.25 is released (https://github.com/open-telemetry/opentelemetry-python/pull/3837)
-        self.assertIn(
-            "CustomError", event.attributes[SpanAttributes.EXCEPTION_TYPE]
+        self.assertEqual(
+            f"{CustomError.__module__}.{CustomError.__qualname__}",
+            event.attributes[EXCEPTION_TYPE],
         )
 
         self.assertEqual(
-            event.attributes[SpanAttributes.EXCEPTION_MESSAGE],
+            event.attributes[EXCEPTION_MESSAGE],
             "The task failed!",
         )