+++ instrumentation/opentelemetry-instrumentation-httpx/src/opentelemetry/instrumentation/httpx/__init__.py
@@ -217,6 +217,97 @@ For example,
 
 will exclude requests such as ``https://site/client/123/info`` and ``https://site/xyz/healthcheck``.
 
+Capture HTTP request and response headers
+*****************************************
+You can configure the agent to capture specified HTTP headers as span attributes, according to the
+`semantic conventions <https://opentelemetry.io/docs/specs/semconv/http/http-spans/#http-client-span>`_.
+
+Request headers
+***************
+To capture HTTP request headers as span attributes, set the environment variable
+``OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_CLIENT_REQUEST`` to a comma delimited list of HTTP header names.
+
+For example using the environment variable,
+::
+
+    export OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_CLIENT_REQUEST="content-type,custom_request_header"
+
+will extract ``content-type`` and ``custom_request_header`` from the request headers and add them as span attributes.
+
+Request header names in HttpX are case-insensitive. So, giving the header name as ``CUStom-Header`` in the environment
+variable will capture the header named ``custom-header``.
+
+Regular expressions may also be used to match multiple headers that correspond to the given pattern.  For example:
+::
+
+    export OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_CLIENT_REQUEST="Accept.*,X-.*"
+
+Would match all request headers that start with ``Accept`` and ``X-``.
+
+To capture all request headers, set ``OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_CLIENT_REQUEST`` to ``".*"``.
+::
+
+    export OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_CLIENT_REQUEST=".*"
+
+The name of the added span attribute will follow the format ``http.request.header.<header_name>`` where ``<header_name>``
+is the normalized HTTP header name (lowercase, with ``-`` replaced by ``_``). The value of the attribute will be a
+single item list containing all the header values.
+
+For example:
+``http.request.header.custom_request_header = ["<value1>", "<value2>"]``
+
+Response headers
+****************
+To capture HTTP response headers as span attributes, set the environment variable
+``OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_CLIENT_RESPONSE`` to a comma delimited list of HTTP header names.
+
+For example using the environment variable,
+::
+
+    export OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_CLIENT_RESPONSE="content-type,custom_response_header"
+
+will extract ``content-type`` and ``custom_response_header`` from the response headers and add them as span attributes.
+
+Response header names in HttpX are case-insensitive. So, giving the header name as ``CUStom-Header`` in the environment
+variable will capture the header named ``custom-header``.
+
+Regular expressions may also be used to match multiple headers that correspond to the given pattern.  For example:
+::
+
+    export OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_CLIENT_RESPONSE="Content.*,X-.*"
+
+Would match all response headers that start with ``Content`` and ``X-``.
+
+To capture all response headers, set ``OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_CLIENT_RESPONSE`` to ``".*"``.
+::
+
+    export OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_CLIENT_RESPONSE=".*"
+
+The name of the added span attribute will follow the format ``http.response.header.<header_name>`` where ``<header_name>``
+is the normalized HTTP header name (lowercase, with ``-`` replaced by ``_``). The value of the attribute will be a
+list containing the header values.
+
+For example:
+``http.response.header.custom_response_header = ["<value1>", "<value2>"]``
+
+Sanitizing headers
+******************
+In order to prevent storing sensitive data such as personally identifiable information (PII), session keys, passwords,
+etc, set the environment variable ``OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_SANITIZE_FIELDS``
+to a comma delimited list of HTTP header names to be sanitized.
+
+Regexes may be used, and all header names will be matched in a case-insensitive manner.
+
+For example using the environment variable,
+::
+
+    export OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_SANITIZE_FIELDS=".*session.*,set-cookie"
+
+will replace the value of headers such as ``session-id`` and ``set-cookie`` with ``[REDACTED]`` in the span.
+
+Note:
+    The environment variable names used to capture HTTP headers are still experimental, and thus are subject to change.
+
 API
 ---
 """
@@ -225,6 +316,7 @@ from __future__ import annotations
 
 import logging
 import typing
+from collections import defaultdict
 from functools import partial
 from inspect import iscoroutinefunction
 from timeit import default_timer
@@ -278,8 +370,15 @@ from opentelemetry.trace import SpanKind, Tracer, TracerProvider, get_tracer
 from opentelemetry.trace.span import Span
 from opentelemetry.trace.status import StatusCode
 from opentelemetry.util.http import (
+    OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_CLIENT_REQUEST,
+    OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_CLIENT_RESPONSE,
+    OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_SANITIZE_FIELDS,
     ExcludeList,
+    get_custom_header_attributes,
+    get_custom_headers,
     get_excluded_urls,
+    normalise_request_header_name,
+    normalise_response_header_name,
     redact_url,
     sanitize_method,
 )
@@ -379,6 +478,32 @@ def _inject_propagation_headers(headers, args, kwargs):
         kwargs["headers"] = _headers.raw
 
 
+def _normalize_headers(
+    headers: httpx.Headers
+    | dict[str, list[str] | str]
+    | list[tuple[bytes, bytes]]
+    | None,
+) -> dict[str, list[str]]:
+    normalized_headers: defaultdict[str, list[str]] = defaultdict(list)
+    if isinstance(headers, httpx.Headers):
+        for key in headers.keys():
+            normalized_headers[key.lower()].extend(
+                headers.get_list(key, split_commas=True)
+            )
+    elif isinstance(headers, dict):
+        for key, value in headers.items():
+            if isinstance(value, list):
+                normalized_headers[key.lower()].extend(value)
+            else:
+                normalized_headers[key.lower()].append(value)
+    elif isinstance(headers, list):
+        for key, value in headers:
+            normalized_headers[key.decode("latin-1").lower()].append(
+                value.decode("latin-1")
+            )
+    return dict(normalized_headers)
+
+
 def _extract_response(
     response: httpx.Response
     | tuple[int, httpx.Headers, httpx.SyncByteStream, dict[str, typing.Any]],
@@ -410,6 +535,9 @@ def _apply_request_client_attributes_to_span(
     url: str | httpx.URL,
     method_original: str,
     semconv: _StabilityMode,
+    headers: httpx.Headers | dict[str, list[str] | str] | None = None,
+    captured_headers: list[str] | None = None,
+    sensitive_headers: list[str] | None = None,
 ):
     url = httpx.URL(url)
     # http semconv transition: http.method -> http.request.method
@@ -429,6 +557,15 @@ def _apply_request_client_attributes_to_span(
         method_original,
         sanitize_method(method_original),
         semconv,
+    )
+
+    span_attributes.update(
+        get_custom_header_attributes(
+            _normalize_headers(headers),
+            captured_headers,
+            sensitive_headers,
+            normalise_request_header_name,
+        )
     )
 
     if _report_old(semconv):
@@ -459,6 +596,9 @@ def _apply_response_client_attributes_to_span(
     status_code: int,
     http_version: str,
     semconv: _StabilityMode,
+    headers: httpx.Headers | dict[str, list[str] | str] | None = None,
+    captured_headers: list[str] | None = None,
+    sensitive_headers: list[str] | None = None,
 ):
     # http semconv transition: http.status_code -> http.response.status_code
     # TODO: use _set_status when it's stable for http clients
@@ -470,6 +610,15 @@ def _apply_response_client_attributes_to_span(
     )
     http_status_code = http_status_to_status_code(status_code)
     span.set_status(http_status_code)
+
+    span.set_attributes(
+        get_custom_header_attributes(
+            _normalize_headers(headers),
+            captured_headers,
+            sensitive_headers,
+            normalise_response_header_name,
+        )
+    )
 
     if http_status_code == StatusCode.ERROR and _report_new(semconv):
         # http semconv transition: new error.type
@@ -573,6 +722,15 @@ class SyncOpenTelemetryTransport(httpx.BaseTransport):
         self._request_hook = request_hook
         self._response_hook = response_hook
         self._excluded_urls = get_excluded_urls("HTTPX")
+        self._captured_request_headers = get_custom_headers(
+            OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_CLIENT_REQUEST
+        )
+        self._captured_response_headers = get_custom_headers(
+            OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_CLIENT_RESPONSE
+        )
+        self._sensitive_headers = get_custom_headers(
+            OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_SANITIZE_FIELDS
+        )
 
     def __enter__(self) -> SyncOpenTelemetryTransport:
         self._transport.__enter__()
@@ -619,6 +777,9 @@ class SyncOpenTelemetryTransport(httpx.BaseTransport):
             url,
             method_original,
             self._sem_conv_opt_in_mode,
+            headers,
+            self._captured_request_headers,
+            self._sensitive_headers,
         )
 
         request_info = RequestInfo(method, url, headers, stream, extensions)
@@ -663,6 +824,9 @@ class SyncOpenTelemetryTransport(httpx.BaseTransport):
                         status_code,
                         http_version,
                         self._sem_conv_opt_in_mode,
+                        headers,
+                        self._captured_response_headers,
+                        self._sensitive_headers,
                     )
                 if callable(self._response_hook):
                     self._response_hook(
@@ -773,6 +937,15 @@ class AsyncOpenTelemetryTransport(httpx.AsyncBaseTransport):
         self._request_hook = request_hook
         self._response_hook = response_hook
         self._excluded_urls = get_excluded_urls("HTTPX")
+        self._captured_request_headers = get_custom_headers(
+            OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_CLIENT_REQUEST
+        )
+        self._captured_response_headers = get_custom_headers(
+            OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_CLIENT_RESPONSE
+        )
+        self._sensitive_headers = get_custom_headers(
+            OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_SANITIZE_FIELDS
+        )
 
     async def __aenter__(self) -> "AsyncOpenTelemetryTransport":
         await self._transport.__aenter__()
@@ -817,6 +990,9 @@ class AsyncOpenTelemetryTransport(httpx.AsyncBaseTransport):
             url,
             method_original,
             self._sem_conv_opt_in_mode,
+            headers,
+            self._captured_request_headers,
+            self._sensitive_headers,
         )
 
         request_info = RequestInfo(method, url, headers, stream, extensions)
@@ -863,6 +1039,9 @@ class AsyncOpenTelemetryTransport(httpx.AsyncBaseTransport):
                         status_code,
                         http_version,
                         self._sem_conv_opt_in_mode,
+                        headers,
+                        self._captured_response_headers,
+                        self._sensitive_headers,
                     )
 
                 if callable(self._response_hook):
@@ -923,6 +1102,7 @@ class HTTPXClientInstrumentor(BaseInstrumentor):
     def instrumentation_dependencies(self) -> typing.Collection[str]:
         return _instruments
 
+    # pylint: disable=too-many-locals
     def _instrument(self, **kwargs: typing.Any):
         """Instruments httpx Client and AsyncClient
 
@@ -954,6 +1134,15 @@ class HTTPXClientInstrumentor(BaseInstrumentor):
             else None
         )
         excluded_urls = get_excluded_urls("HTTPX")
+        captured_request_headers: list[str] = get_custom_headers(
+            OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_CLIENT_REQUEST
+        )
+        captured_response_headers: list[str] = get_custom_headers(
+            OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_CLIENT_RESPONSE
+        )
+        sensitive_headers: list[str] = get_custom_headers(
+            OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_SANITIZE_FIELDS
+        )
 
         _OpenTelemetrySemanticConventionStability._initialize()
         sem_conv_opt_in_mode = _OpenTelemetrySemanticConventionStability._get_opentelemetry_stability_opt_in_mode(
@@ -1003,6 +1192,9 @@ class HTTPXClientInstrumentor(BaseInstrumentor):
                 request_hook=request_hook,
                 response_hook=response_hook,
                 excluded_urls=excluded_urls,
+                captured_request_headers=captured_request_headers,
+                captured_response_headers=captured_response_headers,
+                sensitive_headers=sensitive_headers,
             ),
         )
         wrap_function_wrapper(
@@ -1017,6 +1209,9 @@ class HTTPXClientInstrumentor(BaseInstrumentor):
                 async_request_hook=async_request_hook,
                 async_response_hook=async_response_hook,
                 excluded_urls=excluded_urls,
+                captured_request_headers=captured_request_headers,
+                captured_response_headers=captured_response_headers,
+                sensitive_headers=sensitive_headers,
             ),
         )
 
@@ -1037,6 +1232,9 @@ class HTTPXClientInstrumentor(BaseInstrumentor):
         request_hook: RequestHook,
         response_hook: ResponseHook,
         excluded_urls: ExcludeList | None,
+        captured_request_headers: list[str] | None = None,
+        captured_response_headers: list[str] | None = None,
+        sensitive_headers: list[str] | None = None,
     ):
         if not is_http_instrumentation_enabled():
             return wrapped(*args, **kwargs)
@@ -1059,6 +1257,9 @@ class HTTPXClientInstrumentor(BaseInstrumentor):
             url,
             method_original,
             sem_conv_opt_in_mode,
+            headers,
+            captured_request_headers,
+            sensitive_headers,
         )
 
         request_info = RequestInfo(method, url, headers, stream, extensions)
@@ -1103,7 +1304,11 @@ class HTTPXClientInstrumentor(BaseInstrumentor):
                         status_code,
                         http_version,
                         sem_conv_opt_in_mode,
+                        headers,
+                        captured_response_headers,
+                        sensitive_headers,
                     )
+
                 if callable(response_hook):
                     response_hook(
                         span,
@@ -1158,6 +1363,9 @@ class HTTPXClientInstrumentor(BaseInstrumentor):
         async_request_hook: AsyncRequestHook,
         async_response_hook: AsyncResponseHook,
         excluded_urls: ExcludeList | None,
+        captured_request_headers: typing.Optional[list[str]] = None,
+        captured_response_headers: typing.Optional[list[str]] = None,
+        sensitive_headers: typing.Optional[list[str]] = None,
     ):
         if not is_http_instrumentation_enabled():
             return await wrapped(*args, **kwargs)
@@ -1180,6 +1388,9 @@ class HTTPXClientInstrumentor(BaseInstrumentor):
             url,
             method_original,
             sem_conv_opt_in_mode,
+            headers,
+            captured_request_headers,
+            sensitive_headers,
         )
 
         request_info = RequestInfo(method, url, headers, stream, extensions)
@@ -1224,6 +1435,9 @@ class HTTPXClientInstrumentor(BaseInstrumentor):
                         status_code,
                         http_version,
                         sem_conv_opt_in_mode,
+                        headers,
+                        captured_response_headers,
+                        sensitive_headers,
                     )
 
                 if callable(async_response_hook):
@@ -1341,6 +1555,15 @@ class HTTPXClientInstrumentor(BaseInstrumentor):
             async_response_hook = None
 
         excluded_urls = get_excluded_urls("HTTPX")
+        captured_request_headers: list[str] = get_custom_headers(
+            OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_CLIENT_REQUEST
+        )
+        captured_response_headers: list[str] = get_custom_headers(
+            OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_CLIENT_RESPONSE
+        )
+        sensitive_headers: list[str] = get_custom_headers(
+            OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_SANITIZE_FIELDS
+        )
 
         if hasattr(client._transport, "handle_request"):
             wrap_function_wrapper(
@@ -1355,6 +1578,9 @@ class HTTPXClientInstrumentor(BaseInstrumentor):
                     request_hook=request_hook,
                     response_hook=response_hook,
                     excluded_urls=excluded_urls,
+                    captured_request_headers=captured_request_headers,
+                    captured_response_headers=captured_response_headers,
+                    sensitive_headers=sensitive_headers,
                 ),
             )
             for transport in client._mounts.values():
@@ -1371,6 +1597,9 @@ class HTTPXClientInstrumentor(BaseInstrumentor):
                             request_hook=request_hook,
                             response_hook=response_hook,
                             excluded_urls=excluded_urls,
+                            captured_request_headers=captured_request_headers,
+                            captured_response_headers=captured_response_headers,
+                            sensitive_headers=sensitive_headers,
                         ),
                     )
             client._is_instrumented_by_opentelemetry = True
@@ -1387,6 +1616,9 @@ class HTTPXClientInstrumentor(BaseInstrumentor):
                     async_request_hook=async_request_hook,
                     async_response_hook=async_response_hook,
                     excluded_urls=excluded_urls,
+                    captured_request_headers=captured_request_headers,
+                    captured_response_headers=captured_response_headers,
+                    sensitive_headers=sensitive_headers,
                 ),
             )
             for transport in client._mounts.values():
@@ -1403,6 +1635,9 @@ class HTTPXClientInstrumentor(BaseInstrumentor):
                             async_request_hook=async_request_hook,
                             async_response_hook=async_response_hook,
                             excluded_urls=excluded_urls,
+                            captured_request_headers=captured_request_headers,
+                            captured_response_headers=captured_response_headers,
+                            sensitive_headers=sensitive_headers,
                         ),
                     )
             client._is_instrumented_by_opentelemetry = True